fix doctest error

+++ sklearn/cluster/spectral.py
@@ -386,7 +386,7 @@ class SpectralClustering(BaseEstimator, ClusterMixin):
     ...         assign_labels="discretize",
     ...         random_state=0).fit(X)
     >>> clustering.labels_
-    array([1, 1, 1, 0, 0, 0], dtype=int64)
+    array([1, 1, 1, 0, 0, 0])
     >>> clustering # doctest: +NORMALIZE_WHITESPACE
     SpectralClustering(affinity='rbf', assign_labels='discretize', coef0=1,
               degree=3, eigen_solver=None, eigen_tol=0.0, gamma=1.0,
